+++ src/osm_easy_api/data_classes/relation.py
@@ -1,8 +1,13 @@
 from dataclasses import dataclass, field
 from typing import NamedTuple
+from copy import copy
 
 from ..data_classes.osm_object_primitive import osm_object_primitive
 from ..data_classes import Node, Way
+
+
+_RELATION_DICTIONARY_TYPE = dict[str, str | list["_MEMBER_DICTIONARY_TYPE"]]
+_MEMBER_DICTIONARY_TYPE = dict[str, str | dict[str, str | list[dict[str, str]]] | _RELATION_DICTIONARY_TYPE]
 
 @dataclass
 class Relation(osm_object_primitive):
@@ -19,5 +24,35 @@ class Relation(osm_object_primitive):
             for member, role in self.members:
                 element.appendChild(member._to_xml(changeset_id=changeset_id, member_version=True, role=role))
             return element
+        
+    def to_dict(self) -> dict[str, str | list[_MEMBER_DICTIONARY_TYPE]]:
+        super_dict: dict[str, str | list[_MEMBER_DICTIONARY_TYPE]] = super().to_dict() # type: ignore
+        members: list[_MEMBER_DICTIONARY_TYPE] = []
+        for member in self.members:
+            members.append(member.to_dict())
+        super_dict["members"] = members
+        return super_dict
+    
+    @classmethod
+    def from_dict(cls, dict: dict[str, str | list[_MEMBER_DICTIONARY_TYPE]]):
+        temp_dict = copy(dict)
+        relation = super().from_dict(temp_dict) # type: ignore
+        members = copy(dict["members"])
+        relation.members.clear()
+        for member in members:
+            relation.members.append(Member.from_dict(member))
+        return relation
 
-Member = NamedTuple("Member", [("element", Node | Way | Relation), ("role", str)])
+class Member(NamedTuple("Member", [("element", Node | Way | Relation), ("role", str)])):
+    def to_dict(self) -> _MEMBER_DICTIONARY_TYPE:
+        return {"element": self.element.to_dict(), "role": self.role, "type": "Member"}
+    
+    @classmethod
+    def from_dict(cls, dict) -> "Member":
+        if not dict.get("type"): 
+            raise ValueError("No type key in the dictionary!")
+        match dict["element"]["type"]:
+            case "Node": return cls(Node.from_dict(dict["element"]), dict["role"])
+            case "Way": return cls(Way.from_dict(dict["element"]), dict["role"])
+            case "Relation": return cls(Relation.from_dict(dict["element"]), dict["role"])
+            case _: raise ValueError(f'{dict["type"]} not supported for relation member.')